Refactor cacher.ListerWatcher code structure
