+++ staging/src/k8s.io/apiserver/pkg/storage/tests/cacher_test.go
@@ -557,89 +557,6 @@ func TestEmptyWatchEventCache(t *testing.T) {
 	}
 }
 
-func TestCacherListerWatcher(t *testing.T) {
-	prefix := "pods"
-	fn := func() runtime.Object { return &example.PodList{} }
-	server, store := newEtcdTestStorage(t, prefix, true)
-	defer server.Terminate(t)
-
-	podFoo := makeTestPod("foo")
-	podBar := makeTestPod("bar")
-	podBaz := makeTestPod("baz")
-
-	_ = updatePod(t, store, podFoo, nil)
-	_ = updatePod(t, store, podBar, nil)
-	_ = updatePod(t, store, podBaz, nil)
-
-	lw := cacherstorage.NewCacherListerWatcher(store, prefix, fn)
-
-	obj, err := lw.List(metav1.ListOptions{})
-	if err != nil {
-		t.Fatalf("List failed: %v", err)
-	}
-	pl, ok := obj.(*example.PodList)
-	if !ok {
-		t.Fatalf("Expected PodList but got %v", pl)
-	}
-	if len(pl.Items) != 3 {
-		t.Errorf("Expected PodList of length 3 but got %d", len(pl.Items))
-	}
-}
-
-func TestCacherListerWatcherPagination(t *testing.T) {
-	prefix := "pods"
-	fn := func() runtime.Object { return &example.PodList{} }
-	server, store := newEtcdTestStorage(t, prefix, true)
-	defer server.Terminate(t)
-
-	podFoo := makeTestPod("foo")
-	podBar := makeTestPod("bar")
-	podBaz := makeTestPod("baz")
-
-	_ = updatePod(t, store, podFoo, nil)
-	_ = updatePod(t, store, podBar, nil)
-	_ = updatePod(t, store, podBaz, nil)
-
-	lw := cacherstorage.NewCacherListerWatcher(store, prefix, fn)
-
-	obj1, err := lw.List(metav1.ListOptions{Limit: 2})
-	if err != nil {
-		t.Fatalf("List failed: %v", err)
-	}
-	limit1, ok := obj1.(*example.PodList)
-	if !ok {
-		t.Fatalf("Expected PodList but got %v", limit1)
-	}
-	if len(limit1.Items) != 2 {
-		t.Errorf("Expected PodList of length 2 but got %d", len(limit1.Items))
-	}
-	if limit1.Continue == "" {
-		t.Errorf("Expected list to have Continue but got none")
-	}
-	obj2, err := lw.List(metav1.ListOptions{Limit: 2, Continue: limit1.Continue})
-	if err != nil {
-		t.Fatalf("List failed: %v", err)
-	}
-	limit2, ok := obj2.(*example.PodList)
-	if !ok {
-		t.Fatalf("Expected PodList but got %v", limit2)
-	}
-	if limit2.Continue != "" {
-		t.Errorf("Expected list not to have Continue, but got %s", limit1.Continue)
-	}
-
-	if limit1.Items[0].Name != podBar.Name {
-		t.Errorf("Expected list1.Items[0] to be %s but got %s", podBar.Name, limit1.Items[0].Name)
-	}
-	if limit1.Items[1].Name != podBaz.Name {
-		t.Errorf("Expected list1.Items[1] to be %s but got %s", podBaz.Name, limit1.Items[1].Name)
-	}
-	if limit2.Items[0].Name != podFoo.Name {
-		t.Errorf("Expected list2.Items[0] to be %s but got %s", podFoo.Name, limit2.Items[0].Name)
-	}
-
-}
-
 func TestWatchDispatchBookmarkEvents(t *testing.T) {
 	ctx, cacher, terminate := testSetup(t)
 	t.Cleanup(terminate)
